# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Digital HubX API.",
        "title": "Digital HubX API",
        "contact": {
            "name": "API Support (Tanapong R)",
            "url": "mailto:zTanapongR@pttep.com",
            "email": "zTanapongR@pttep.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/hubx-service",
    "paths": {
...
        "/admin/class": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an admin to create a new class by providing title, description, category, tier, and other details.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Class"
                ],
                "summary": "Create a new class",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Class Description",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Class Cover Image",
                        "name": "cover_image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Class Category ID",
                        "name": "class_category_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Class Tier",
                        "name": "class_tier",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Class Level",
                        "name": "class_level",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful\" example:{\"message\":\"Class created successfully\",\"status\":200,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input\" example:{\"message\":\"Invalid input\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/class-category": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an admin to create a new class category.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Class"
                ],
                "summary": "Create a new class category",
                "parameters": [
                    {
                        "description": "Create Class Category Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category created successfully\" example:{\"message\":\"Category created successfully\",\"status\":200,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input\" example:{\"message\":\"The input data is invalid\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/admin/class/{class_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft deletes a class by marking it as removed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Class"
                ],
                "summary": "Delete a class",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful\" example:{\"message\":\"Class deleted successfully\",\"status\":200,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Invalid input\" example:{\"message\":\"Class not found\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/class/{class_id}/cover-image": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an admin to update the cover image of a class.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "Update class cover image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "New Cover Image",
                        "name": "new_cover_image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful\" example:{\"message\":\"Class cover image updated successfully\",\"status\":200,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input\" example:{\"message\":\"Invalid input data\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Invalid input\" example:{\"message\":\"Class not found\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/class/{class_id}/toggle-enable-question": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Enables or disables question functionality for a specific class.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "Toggle EnableQuestion status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful\" example:{\"message\":\"EnableQuestion status updated\",\"status\":200,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Invalid input\" example:{\"message\":\"Class not found\",\"status\":400,\"details\":{\"field\":\"error description\"}}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/class": {
            "get": {
                "description": "Fetch all classes with optional filters like tier, keyword, and category.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "Get all classes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default: 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Class Tier",
                        "name": "class_tier",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by class level",
                        "name": "class_level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by class category",
                        "name": "class_category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful\" example:{\"message\":\"Class created successfully\",\"status\":200,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error\" example:{\"message\":\"Internal Server Error\",\"status\":500,\"details\":null}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/class/{class_id}": {
            "get": {
                "description": "Fetch the details of a specific class by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "Get class by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Class ID",
                        "name": "class_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class details",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClassRes"
                        }
                    },
                    "404": {
                        "description": "Class not found",
                        "schema": {
                            "$ref": "#/definitions/response.MsgResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.MsgResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateClassRes": {
            "description": "Response body after creating a class.",
            "type": "object",
            "properties": {
                "class_category_id": {
                    "type": "string"
                },
                "class_level": {
                    "type": "integer"
                },
                "class_tier": {
                    "$ref": "#/definitions/models.ClassTier"
                },
                "cover_image": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_remove": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.ClassTier": {
            "type": "string",
            "enum": [
                "essential",
                "literacy",
                "mastery"
            ],
            "x-enum-varnames": [
                "Essential",
                "Literacy",
                "Mastery"
            ]
        },
...
        "response.MsgResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Additional details (optional)",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "{\"field\"": " \"error detail\"}"
                    }
                },
                "message": {
                    "description": "The response message",
                    "type": "string",
                    "example": "Operation successful"
                },
                "status": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 200
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "This security definition is used for authenticating",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}